requirements.txt
pydantic~=2.8.2
python-dotenv~=1.0.1
fastapi~=0.112.1
SQLAlchemy~=2.0.32
starlette~=0.38.2
uvicorn~=0.30.6
PyMySQL~=1.1.1
cryptography~=43.0.0